Refactor 'attach --pool' a bit
Now it updates progress as it tries to attach to
each pool. It also shows which pool id resulted
in which product.

Removed some deeply nested code.

Add a read_pool_ids() that takes a fileobject, to
make it easier to test.

Note this could allow the command to continue on
pool attach errors, instead of exiting at the
first failure.

Most of the "attach a pool" code is ready to be
seperated into a seperate file/module now.